Modularize and utilize components
