+++ wails/wallet/frontend/src/views/CreateWallet.svelte
@@ -1,40 +1,6 @@
 <script>
   import { CreateWallet } from '../../wailsjs/go/main/App'    
-
-  let passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).{12,}$/;
-  function isPasswordValid(password) {
-    return passwordRegex.test(password);
-  }
-
-  function checkPasswordStrength() {
-    const password = document.getElementById('wallet-password').value;
-    const passwordStrengthContainer = document.getElementById('password-strength');
-    const isPasswordStrong = isPasswordValid(password);
-    passwordStrengthContainer.style.display = 'block';
-    if (isPasswordStrong) {
-      passwordStrengthContainer.textContent = 'Password strength: Strong';
-     } else {
-        passwordStrengthContainer.textContent = 'Password strength: Weak';
-      }
-    }
-
-  function checkPasswordsMatch() {
-    const password = document.getElementById('wallet-password').value;
-    const confirmPassword = document.getElementById('confirm-password').value;
-    const isPasswordStrong = isPasswordValid(password);
-    const createWalletButton = document.getElementById('create-wallet-button');
-    const passwordMatchContainer = document.getElementById('password-match');
-    passwordMatchContainer.style.display = 'block';
-    if (password !== confirmPassword) {
-      passwordMatchContainer.textContent = 'Passwords do not match';
-    } else {
-      passwordMatchContainer.textContent = '';
-      if (isPasswordStrong) {
-        createWalletButton.disabled = false;
-        createWalletButton.style.backgroundColor = 'black';
-      }
-    }
-  } 
+  import CreatePassword from '../components/CreatePassword.svelte';
 
   function createWallet() {
     const password = document.getElementById('wallet-password').value;
@@ -48,31 +14,15 @@
   }
   
   </script>
-  
   <main>
-    <div class="header-content">
-      <h2>Create Wallet</h2>
-      <h3>Create Password</h3>
-      <p>This password will unlock your wallet on this application.</p>
-    </div>
-  
-    <form class="form-container">
-      <div class="input-group">
-        <label for="password">New password (16 characters min, at least one character should be in lowercase, one in uppercase and one number)</label>
-        <input type="password" id="wallet-password" on:input={checkPasswordStrength} />
-        <p id="password-strength" class="validation-label" ></p>
-      </div>
-      <div class="input-group">
-        <label for="confirm-password">Confirm password</label>
-        <input type="password" id="confirm-password" on:input={checkPasswordsMatch} />
-        <p id="password-match" class="validation-label" ></p>
-      </div>
-      <button id="create-wallet-button" type="button" disabled=true on:click={createWallet}>Create a new wallet</button>
-    </form>
+    <CreatePassword
+      handleClick={createWallet}
+      walletLabel="Create Wallet"
+    />
   </main>
-  
-  <style>
-    main {
+
+<style>
+  main {
     font-family: "Nunito", sans-serif;
     color: #333; 
     background-color: #f9f9f9; 
@@ -83,145 +33,4 @@
     align-items: center;
   }
 
-  .header-content {
-    text-align: center;
-    margin-bottom: 20px;
-  }
-
-  h2 {
-    font-size: 2rem;
-    color: #0066cc; 
-    margin-bottom: 10px;
-  }
-
-  h3 {
-    font-size: 1.5rem;
-    color: #333;
-    margin-bottom: 10px;
-  }
-
-  p {
-    font-size: 1rem;
-    color: #555; 
-    margin-bottom: 20px;
-  }
-
-  /* Form styles */
-  .form-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    align-content: center;
-    gap: 20px;
-    width: 100%;
-    max-width: 400px;
-    background-color: #fff;
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1); 
-  }
-
-  .input-group {
-    display: flex;
-    flex-direction: column;
-    gap: 2px;
-    align-items: center;
-  }
-
-  label {
-    font-size: 1rem;
-    font-weight: bold;
-    color: #222;
-  }
-
-  input {
-    padding: 4px;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    max-width: 190px;
-    color: #333; 
-    background-color: #fefefe; 
-  }
-
-  input:focus {
-    outline: none;
-    border-color: #0066cc; 
-    box-shadow: 0 0 3px #0066cc;
-  }
-
-  #create-wallet-button {
-    text-align: center;
-    color: white;
-    background-color: #0066cc;
-    font-family: "Nunito", sans-serif;
-    font-weight: bold;
-    width: 100%;
-    height: 25px;
-    border-radius: 25px;
-    cursor: pointer;
-    border: none;
-    transition: background-color 0.3s ease;
-  }
-
-  #create-wallet-button:hover:not(:disabled) {
-    background-color: #004c99; 
-  }
-
-  #create-wallet-button:disabled {
-    background-color: #ccc; 
-    color: #666;
-    cursor: not-allowed;
-  }
-
-  /* Validation messages */
-  .validation-label {
-    font-size: 0.9rem;
-    color: red; 
-    display: none; 
-  }
-
-  /* Responsive styles */
-  @media (max-width: 768px) {
-    main {
-      padding: 15px;
-    }
-
-    h2 {
-      font-size: 1.8rem;
-    }
-
-    h3 {
-      font-size: 1.3rem;
-    }
-
-    .form-container {
-      padding: 15px;
-    }
-
-    #create-wallet-button {
-      height: 40px;
-    }
-  }
-
-  @media (max-width: 480px) {
-    h2 {
-      font-size: 1.5rem;
-    }
-
-    h3 {
-      font-size: 1.2rem;
-    }
-
-    label,
-    input,
-    #create-wallet-button {
-      font-size: 0.9rem;
-    }
-
-    .form-container {
-      gap: 15px;
-    }
-  }
-  </style>
-  
+</style>
